Refactor client cryptography code.
